Reimplement EyeTracking based on np.ndarray
Reimplement EyeTracking and TrialEyeTracking using a `np.ndarray` `data`
attribute instead of a `pd.DataFrame`. This makes the implementation
more consistent with `Fluorescence` classes and simplifies the
implementation of methods for subsetting and computing summary
statistics.

Additionally, the `dict` of `ndarray`s provided by `data` is made
read-only via a `MappingProxyType` of a protected `_data` attribute.
This should prevent some accidental modification of `data` (however, for
now the arrays contained in the read-only dict are themselves mutable,
so it is still possible to modify the contents of `data` members). This
required the reimplementation of `__copy__` and `__deepcopy__` for
EyeTracking.